Add feature freeze workflow
Behaviour:
For each new PR (or edited description), we look to the changelog and
look if it mentions a lint never seen before in the code. If it's the first
mention of the lint in the codebase, we flag it as a new lint, add the label
and add it to the milestone.